Crate mioco

Coroutine-based handler library for mio

Using coroutines, an event-based mio model can be simplified to a set of routines, seamlessly scheduled on demand in userspace.

With mioco a coroutines can be used to simplify writing asynchronous io handling in synchronous fashion.

See examples/echo.rs for an example TCP echo server.

Structs

LastEvent
Last Event

MiocoHandle
Mioco Handle

TypedIOSource
mioco wrapper over raw mio IO structure

Enums

RW
Read/Write/Both

Traits

Evented
mioco can work on any type implementing this trait

Functions

start
Start mioco handling