Function mioco::select_wait

pub fn select_wait() -> Event

Block the current coroutine waiting for an event.

Use select! macro instead.

Warning: Mioco can't guarantee that the returned EventSource will not block when actually attempting to read or write. You must use try_read and try_write instead to handle spurious wakeups.

The returned value contains event type and the id of the EventSource. See EventSourceId.