Module mioco::mio

Modules

buf

prelude

tcp

udp

unix

util
Utilities for non-blocking IO programs

Structs

EventLoop
Single threaded IO event loop.

EventLoopConfig
Configure EventLoop runtime details

EventSet

Io

Ipv4Addr
Representation of an IPv4 address.

Ipv6Addr
Representation of an IPv6 address.

Poll

PollOpt

Selector

Sender
Sends messages to the EventLoop from other threads.

Timeout

TimerError

Token

Enums

IpAddr
An IP address, either a IPv4 or IPv6 address.

NotifyError

Traits

Buf
A trait for values that provide sequential read access to bytes.

Evented
A value that may be registered with an EventLoop

Handler

MutBuf
A trait for values that provide sequential write access to bytes.

TryAccept

TryRead

TryWrite

Type Definitions

TimerResult

Implementations