Trait mioco::sched::SchedulerThread

pub trait SchedulerThread {
    fn spawned(&mut self, event_loop: &mut EventLoop<Handler>, coroutine_ctrl: CoroutineControl);
    fn ready(&mut self, event_loop: &mut EventLoop<Handler>, coroutine_ctrl: CoroutineControl);

    fn tick(&mut self, _event_loop: &mut EventLoop<Handler>) { ... }
    fn timeout(&mut self) -> Option<u64> { ... }
}

Scheduler thread

Required Methods

fn spawned(&mut self, event_loop: &mut EventLoop<Handler>, coroutine_ctrl: CoroutineControl)

New coroutine was spawned.

This can be used to run it immediately (see CoroutineControl::resume()), save it to be started later, or migrate it to different thread immediately (see CoroutineControl::migrate()).

Dropping coroutine_ctrl means the corresponding coroutine will be killed.

fn ready(&mut self, event_loop: &mut EventLoop<Handler>, coroutine_ctrl: CoroutineControl)

A Coroutine became ready.

coroutine_ctrl is a control reference to the Coroutine that became ready (to be resumed). It can be resumed immediately, or stored somewhere to be resumed later.

Dropping coroutine_ctrl means the corresponding coroutine will be killed.

Provided Methods

fn tick(&mut self, _event_loop: &mut EventLoop<Handler>)

Mio's tick have completed.

Mio signals events in batches, after which a tick is signaled.

All events events have been processed and all unblocked coroutines signaled with SchedulerThread::ready().

After returning from this function, mioco will let mio process a new batch of events.

fn timeout(&mut self) -> Option<u64>

Set the maximum time till the next tick.

timeout will be called after each tick when all ready and spawned coroutines have been delivered to the scheduler. It can be used to force maximum time before next tick in case of no earlier events.

Returning None means no timeout. Some(time) is a time in ms.