Struct mioco::Builder

pub struct Builder {
    // some fields omitted
}

mioco coroutine builder

Create one with new, then use wrap_io on io that you are going to use in the coroutine that you spawn with start.

Methods

impl Builder

fn new() -> Builder

Create new Coroutine builder

fn wrap_io<H, T: 'static>(&mut self, event_loop: &mut EventLoop<H>, io: T, token: Token) -> ExternalHandle where H: Handler, T: Evented

Register mio's io to be used within mioco coroutine

Consumes the io, returns a Handle to a mio wrapper over it.

fn start<F, H>(self, f: F, event_loop: &mut EventLoop<H>) where F: FnOnce(&mut CoroutineHandle) + Send + 'static, H: Handler

Create a mioco coroutine handler

f is routine handling connection. It should not use any blocking operations, and use it's argument for all IO with it's peer