Struct mioco::Coroutine

pub struct Coroutine {
    // some fields omitted
}

Coroutine handling single mio connection

Create with new, and call readable and writeable from main mio main Handler.

Methods

impl Coroutine

fn new<F, H>(stream: TcpStream, event_loop: &mut EventLoop<H>, token: Token, f: F) -> Coroutine where F: FnOnce(&mut IOHandle) + Send + 'static, H: Handler

Create a mioco coroutine handler

f is routine handling connection. It should not use any blocking operations, and use it's argument for all IO with it's peer

fn is_finished(&self) -> bool

Is this mioco coroutine ready to reclaim?

fn readable<H>(&mut self, event_loop: &mut EventLoop<H>, token: Token, hint: ReadHint) where H: Handler

Readable event handler

This is based on mio's readable method in Handler trait.

fn writable<H>(&mut self, event_loop: &mut EventLoop<H>, token: Token) where H: Handler

Readable event handler

This is based on mio's writeable method in Handler trait.