Struct mioco::CoroutineControl

pub struct CoroutineControl {
    // some fields omitted
}

Coroutine control block

Through this interface Coroutine can be resumed and migrated.

Methods

impl CoroutineControl

fn resume(self, event_loop: &mut EventLoop<Handler>)

Resume Coroutine

Panics if Coroutine is not in Ready state.

fn migrate(self, event_loop: &mut EventLoop<Handler>, thread_id: usize)

Migrate to a different thread

Move this Coroutine to be executed on a SchedulerThread for a given thread_id.

Will panic if thread_id is not valid.

fn is_yielding(&self) -> bool

Is this Coroutine ready after yield_now()?

fn get_userdata<'a, T: Any>(&'a self) -> Option<&'a T>

Gets a reference to the user data set through set_userdata. Returns None if T does not match or if no data was set