Struct mioco::EventSource

pub struct EventSource<T> {
    // some fields omitted
}

Event source inside a coroutine

Event sources are a core of Mioco. Mioco coroutines use them to handle IO in a blocking fashion.

They come in different flavours and can be created from native mio types by wrapping withing a coroutine with MiocoHandle::wrap() or type-specific constructors like mailbox() or MiocoHandle::timeout().

Methods

impl<T> EventSource<T> where T: Reflect + 'static

fn with_raw<F, R>(&self, f: F) -> R where F: Fn(&T) -> R

Access raw mio type

fn with_raw_mut<F, R>(&mut self, f: F) -> R where F: Fn(&mut T) -> R

Access mutable raw mio type

fn id(&self) -> EventSourceId

Index identificator of a EventSource

impl<T> EventSource<T> where T: TryAccept + Reflect + 'static

fn accept(&self) -> Result<T::Output>

Block on accept

impl<T> EventSource<MailboxInnerEnd<T>> where T: Reflect + 'static

fn recv(&mut self) -> Result<T>

Receive T sent using corresponding MailboxOuterEnd::send().

Will block coroutine if no elements are available.

impl EventSource<Timer>

fn read(&self) -> Result<()>

Read a timer to block on it until it is done.

fn set_timeout(&mut self, delay_ms: i64)

Set timeout for the timer

The timeout counts from the time set_timer is called.