Struct mioco::Handler

pub struct Handler {
    // some fields omitted
}

Mioco event loop Handler

Registered in mio_orig::EventLoop and implementing mio_orig::Handler. This struct is quite internal so you should not have to worry about it.

Methods

impl Handler

fn new(shared: RcHandlerShared, scheduler: Box<SchedulerThread>) -> Self

Create a Handler.

fn shared(&self) -> &RcHandlerShared

Data shared between Handler and Coroutines belonging to it

fn deliver_to_scheduler(&mut self, event_loop: &mut EventLoop<Self>)

To prevent recursion, all the newly spawned or newly made ready Coroutines are delivered to scheduler here.