Struct mioco::IOHandle

pub struct IOHandle {
    // some fields omitted
}

mioco wrapper over io associated with a given coroutine.

To be used to trigger events inside mioco coroutine. Create using Builder::io_wrap.

It implements readable and writable, corresponding to original mio::Handler methods. Call these from respective mio::Handler.

Methods

impl IOHandle

fn is_finished(&self) -> bool

Is this coroutine finishd

fn ready<H>(&mut self, event_loop: &mut EventLoop<H>, token: Token, events: EventSet) where H: Handler

Readable event handler

This corresponds to mio::Hnalder::readable().

fn deregister<H>(&mut self, event_loop: &mut EventLoop<H>) where H: Handler

Deregister IO from event loop

Must be called for every IO, after is_finished was detected