Struct mioco::MiocoHandle

pub struct MiocoHandle<'a> {
    // some fields omitted
}

Mioco instance handle

Every coroutine gets an instance of this struct as it's argument. Use it from within coroutines to call mioco functionality.

Methods

impl<'a> MiocoHandle<'a>

fn spawn<F>(&mut self, f: F) -> CoroutineHandle where F: FnOnce(&mut MiocoHandle) -> Result<()> + Send + 'static

Create a mioco coroutine handler

f is routine handling connection. It must not use any real blocking-IO operations, only mioco provided types (EventSource) and MiocoHandle functions. Otherwise mioco cooperative scheduling can block on real blocking-IO which defeats using mioco.

fn sync<'b, F, R>(&mut self, f: F) -> R where F: FnMut() -> R + 'b

Execute a block of synchronous operations

This will execute a block of synchronous operations without blocking cooperative coroutine scheduling. This is done by offloading the synchronous operations to a separate thread, a notifying the coroutine when the result is available.

TODO: find some wise people to confirm if this is sound TODO: use threadpool to prevent potential system starvation?

fn wrap<T: 'static>(&mut self, raw_io: T) -> EventSource<T> where T: Evented

Register mio's native io type to be used within mioco coroutine

Consumes the io, returns a mioco wrapper over it. Use this wrapped IO to perform IO.

fn unwrap<T: 'static>(&mut self, io: EventSource<T>) -> T where T: Evented

Deregister mio's native io type from owning mioco coroutine

Consumes the wrapped io, returns the original io.

This function is useful, when EventSource<T> was used in one coroutine, and then needs to be moved to another.

fn thread_num(&self) -> usize

Get number of threads of the Mioco instance that coroutine is running in.

This is useful for load balancing: spawning as many coroutines as there is handling threads that can run them.

fn timer(&mut self) -> &mut EventSource<Timer>

Get mutable reference to a timer source io for this coroutine

Each coroutine has one internal Timer source, that will become readable when it's timeout (see set_timer() ) expire.

fn sleep(&mut self, time_ms: i64)

Block coroutine for a given time

Warning: The precision of this call (and other timer() like functionality) is limited by mio event loop settings. Any small value of time_ms will effectively be rounded up to mio::EventLoop::timer_tick_ms().

fn yield_now(&mut self)

Yield coroutine execution

Coroutine can yield execution without blocking on anything particular to allow scheduler to run other coroutines before resuming execution of the current one.

For this to be effective, custom scheduler must be implemented. See trait Scheduler.

Note: named yield_now as yield is reserved word.

fn select(&mut self) -> Event

Wait till an event is ready

Warning: Mioco can't guarantee that the returned EventSource will not block when actually attempting to read or write. You must use try_read and try_write instead.

The returned value contains event type and the id of the EventSource. See EventSource::id().

fn select_read(&mut self) -> Event

Wait till a read event is ready

See MiocoHandle::select.

fn select_write(&mut self) -> Event

Wait till a read event is ready.

See MiocoHandle::select.

fn select_from(&mut self, ids: &[EventSourceId]) -> Event

Wait till any event is ready on a set of Handles.

See EventSource::id(). See MiocoHandle::select().

fn select_write_from(&mut self, ids: &[EventSourceId]) -> Event

Wait till write event is ready on a set of Handles.

See MiocoHandle::select_from.

fn select_read_from(&mut self, ids: &[EventSourceId]) -> Event

Wait till read event is ready on a set of Handles.

See MiocoHandle::select_from.