Struct mioco::TypedIOSource

pub struct TypedIOSource<T> {
    // some fields omitted
}

mioco wrapper over raw mio IO structure

Create using MiocoHandle::wrap()

It implements standard library Read and Write and other operations, blocking, switching and resuming handler function to build cooperative scheduling on top of asynchronous operations.

Methods

impl<T> TypedIOSource<T> where T: Reflect + 'static

fn with_raw<F>(&self, f: F) where F: Fn(&T)

Access raw mio type

fn with_raw_mut<F>(&mut self, f: F) where F: Fn(&mut T)

Access mutable raw mio type

impl<T> TypedIOSource<T> where T: TryAccept + Reflect + 'static

fn accept(&self) -> Result<T::Output>

Blocked accept