Function mioco::sync::mpsc::channel

pub fn channel<T>() -> (Sender<T>, Receiver<T>)

Create a channel

Channel can be used to deliver data via MPSC queue.

Channel is modeled after std::sync::mpsc::channel(), only supporting mioco-aware sending and receiving.

When receiving end is outside of coroutine, channel will behave just like std::sync::mpsc::channel().