Struct mioco::sync::mpsc::Sender

pub struct Sender<T> {
    // some fields omitted
}

Channel sending end

Use this inside mioco coroutines or outside of mioco itself to send data asynchronously to the receiving end.

Create with channel()

Methods

impl<T> Sender<T>

fn send(&self, t: T) -> Result<(), SendError<T>>

Deliver T to the other end of the channel.

Channel behaves like a queue.

This is non-blocking operation.