Struct mioco::sync::RwLock

pub struct RwLock<T: ?Sized> {
    // some fields omitted
}

A reader-writer lock

Based on std::sync::RwLock. Calls mioco::yield_now() on contention.

Methods

impl<T> RwLock<T>

fn new(t: T) -> Self

Creates a new instance of an RwLock which is unlocked.

impl<T: ?Sized> RwLock<T>

fn native_lock(&self) -> &RwLock<T>

Get a reference to raw std::sync::RwLock.

Use it to perform operations outside of mioco coroutines.

fn read(&self) -> LockResult<RwLockReadGuard<T>>

Locks this rwlock with shared read access, blocking the current coroutine until it can be acquired.

fn try_read(&self) -> TryLockResult<RwLockReadGuard<T>>

Attempts to acquire this rwlock with shared read access.

fn write(&self) -> LockResult<RwLockWriteGuard<T>>

Locks this rwlock with exclusive write access, blocking the current coroutine until it can be acquired.

fn try_write(&self) -> TryLockResult<RwLockWriteGuard<T>>

Attempts to lock this rwlock with exclusive write access.

fn is_poisoned(&self) -> bool

Determines whether the lock is poisoned.