Crate mioco

Mioco

Scalable, coroutine-based, asynchronous IO handling library for Rust programming language.

Using mioco you can handle asynchronous mio-based IO, using set of synchronous-IO handling functions. Based on mio events mioco will cooperatively schedule your handlers.

You can think of mioco as of Node.js for Rust or green threads on top of mio.

Mioco is a library building on top of mio. Mio API is re-exported as mioco::mio.

# Features:

 * multithreading support; (see `Config::set_thread_num()`)
 * user-provided scheduling; (see `Config::set_scheduler()`);
 * support for all native `mio` types (see `MiocoHandle::wrap()`);
 * timers (see `MiocoHandle::timer()`);
 * mailboxes (see `mailbox()`);
 * coroutine exit notification (see `CoroutineHandle::exit_notificator()`).
 * synchronous operations support (see `MiocoHandle::sync()`).

# Example:

See examples/echo.rs for an example TCP echo server:

extern crate mioco;
extern crate env_logger;

use std::net::SocketAddr;
use std::str::FromStr;
use std::io::{Read, Write};
use mioco::tcp::TcpListener;

const DEFAULT_LISTEN_ADDR : &'static str = "127.0.0.1:5555";

fn listend_addr() -> SocketAddr {
    FromStr::from_str(DEFAULT_LISTEN_ADDR).unwrap()
}

fn main() {
    env_logger::init().unwrap();

    mioco::start(||{
        let addr = listend_addr();

        let listener = TcpListener::bind(&addr).unwrap();

        println!("Starting tcp echo server on {:?}", listener.local_addr().unwrap());

        let listener = listener;

        loop {
            let conn = try!(listener.accept());

            mioco::spawn(|| {
                let mut conn = conn;

                let mut buf = [0u8; 1024 * 16];
                loop {
                    let size = try!(conn.read(&mut buf));
                    if size == 0 {/* eof */ break; }
                    try!(conn.write_all(&mut buf[0..size]))
                }

                Ok(())
            });
        }
    });
}

Modules

mail
Mailboxes

mio
Re-export of some mio symbols.

sync
Useful synchronization primitives

tcp
TCP IO

timer
Timers

udp
UDP IO

unix
Unix sockets IO

Macros

select!

Structs

Config
Mioco builder

Coroutine
Mioco coroutine (a.k.a. mioco handler)

CoroutineControl
Coroutine control block

CoroutineHandle
Handle to spawned coroutine

CoroutineId
Id of a Coroutine used to enumerate them

Event
Event delivered to the coroutine

EventSourceId
Id of an event source used to enumerate them

Handler
Mioco event loop Handler

Mioco
Mioco instance

RW
Read/Write/Both

Select
A Select handling struct

Enums

ExitStatus
Coroutine exit status (value returned or panic)

Message
EventLoop message type

Traits

Evented
Mioco event source.

Scheduler
Coroutine Scheduler

SchedulerThread
Per-thread Scheduler

Functions

sleep
Block coroutine for a given time

spawn
Spawn a mioco coroutine

spawn_ext
Spawn a mioco coroutine

start
Shorthand for creating new Mioco instance and starting it right away.

start_threads
Shorthand for creating new Mioco instance with a fixed number of threads and starting it right away.

sync
Execute a block of synchronous operations

thread_num
Get number of threads of the Mioco instance that coroutine is running in.

timer
Get mutable reference to a timer source io for this coroutine

yield_now
Yield coroutine execution