Crate mioco

Coroutine-based handler library for mio

Using coroutines, an event-based mio model can be simplified to a set of routines, seamlessly scheduled on demand in userspace.

With mioco a coroutines can be used to simplify writing asynchronous io handilng in synchronous fashion.

Structs

Builder
mioco coroutine builder

CoroutineHandle
Coroutine control

ExternalHandle
mioco wrapper over io associated with a given coroutine.

InternalHandle
mioco wrapper over io associated with a given coroutine.

LastEvent
Last Event

Enums

RW
Read/Write/Both

Traits

Evented
mioco can work on any type implementing this trait