Crate mioco

Mioco

Scalable, coroutine-based, asynchronous IO handling library for Rust programming language.

Using mioco you can handle asynchronous mio-based IO, using set of synchronous-IO handling functions. Based on mio events mioco will cooperatively schedule your handlers.

You can think of mioco as of Node.js for Rust or green threads on top of mio.

Mioco is a library building on top of mio. Mio API is re-exported as mioco::mio.

# Features:

 * multithreading support; (see `Config::set_thread_num()`)
 * user-provided scheduling; (see `Config::set_scheduler()`);
 * support for all native `mio` types (see `MiocoHandle::wrap()`);
 * timers (see `MiocoHandle::timer()`);
 * mailboxes (see `mailbox()`);
 * coroutine exit notification (see `CoroutineHandle::exit_notificator()`).

# Example:

See examples/echo.rs for an example TCP echo server:

extern crate mioco;
extern crate env_logger;

use std::net::SocketAddr;
use std::str::FromStr;
use std::io::{Read, Write};
use mioco::mio::tcp::{TcpSocket};

const DEFAULT_LISTEN_ADDR : &'static str = "127.0.0.1:5555";

fn listend_addr() -> SocketAddr {
    FromStr::from_str(DEFAULT_LISTEN_ADDR).unwrap()
}

fn main() {
    env_logger::init().unwrap();

    mioco::start(move |mioco| {
        let addr = listend_addr();

        let sock = try!(TcpSocket::v4());
        try!(sock.bind(&addr));
        let sock = try!(sock.listen(1024));

        println!("Starting tcp echo server on {:?}", sock.local_addr().unwrap());

        let sock = mioco.wrap(sock);

        loop {
            let conn = try!(sock.accept());

            mioco.spawn(move |mioco| {
                let mut conn = mioco.wrap(conn);

                let mut buf = [0u8; 1024 * 16];
                loop {
                    let size = try!(conn.read(&mut buf));
                    if size == 0 {
                        /* eof */
                        break;
                    }
                    try!(conn.write_all(&mut buf[0..size]))
                }

                Ok(())
            });
        }
    });
}

Modules

mio

Structs

Config
Mioco builder

Coroutine
Mioco coroutine (a.k.a. mioco handler)

CoroutineControl
Coroutine control block

CoroutineHandle
Handle to spawned coroutine

CoroutineId
Id of a Coroutine used to enumerate them

Event
Event delivered to the coroutine

EventSource
Event source inside a coroutine

EventSourceId
Id of an event source used to enumerate them

Handler
Mioco event loop Handler

MailboxInnerEnd
Inner Mailbox End

MailboxOuterEnd
Outside Mailbox End

Mioco
Mioco instance

MiocoHandle
Mioco instance handle

Timer
A Timer generating event after a given time

Enums

ExitStatus
Coroutine exit status (value returned or panic)

Message
EventLoop message type

Traits

Evented
mioco can work on any type implementing this trait

Scheduler
Coroutine Scheduler

SchedulerThread
Per-thread Scheduler

Functions

mailbox
Create a Mailbox

start
Shorthand for creating new Mioco instance and starting it right away.

start_threads
Shorthand for creating new Mioco instance with a fixed number of threads and starting it right away.