Struct mioco::MiocoHandle

pub struct MiocoHandle {
    // some fields omitted
}

Mioco Handle

Use this from withing coroutines to perform mioco-provided functionality

Methods

impl MiocoHandle

fn spawn<F>(&self, f: F) where F: FnOnce(&mut MiocoHandle) -> Result<()> + 'static

Create a mioco coroutine handler

f is routine handling connection. It must not use any real blocking-IO operations, only mioco provided types (TypedEventSource) and MiocoHandle functions. Otherwise mioco cooperative scheduling can block on real blocking-IO which defeats using mioco.

fn wrap<T: 'static>(&mut self, io: T) -> TypedEventSource<T> where T: Evented

Register mio's native io type to be used within mioco coroutine

Consumes the io, returns a mioco wrapper over it. Use this wrapped IO to perform IO.

fn timeout(&mut self, delay: u64) -> TypedEventSource<Timer>

Create a new timer which will provide a timeout event after delay milliseconds.

fn sleep(&mut self, delay: u64)

Pause the current co-routine for delay milliseconds.

fn select(&mut self) -> LastEvent

Wait till an event is ready

The returned value contains event type and the index id of the TypedEventSource. See TypedEventSource::index().

fn select_read(&mut self) -> LastEvent

Wait till a read event is ready

See MiocoHandle::select.

fn select_write(&mut self) -> LastEvent

Wait till a read event is ready.

See MiocoHandle::select.

fn select_from(&mut self, indices: &[EventSourceIndex]) -> LastEvent

Wait till any event is ready on a set of Handles.

See TypedEventSource::index(). See MiocoHandle::select().

fn select_write_from(&mut self, indices: &[EventSourceIndex]) -> LastEvent

Wait till write event is ready on a set of Handles.

See MiocoHandle::select_from.

fn select_read_from(&mut self, indices: &[EventSourceIndex]) -> LastEvent

Wait till read event is ready on a set of Handles.

See MiocoHandle::select_from.